+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### **Added**
 
 - added `ray-on-eks`, and `manifests/ray-on-eks` manifests
-- Added a `sagemaker-model-monitoring-module` module with an example of data quality and model quality monitoring of a SageMaker Endpoint.
-- Added an option to enable data capture in the `sagemaker-endpoint-module`.
+- added a `sagemaker-model-monitoring-module` module with an example of data quality, model quality, model bias, and model explainability monitoring of a SageMaker Endpoint
+- added an option to enable data capture in the `sagemaker-endpoint-module`
+- added a `personas` example module to deploy various roles required for an AI/ML project 
 
 ### **Changed**
 - remove explicit module manifest account/region mappings from `fmops-qna-rag`
@@ -22,6 +23,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - changed encryption for each bucket to KMS_MANAGED
 - refactor `airflow-dags` module to use Pydantic
 - fix inputs for `bedrock-finetuning` module not working
+- add `retention-type` argument for the bucket in the `bedrock-finetuning` module
+- fix broken dependencies for `examples/airflow-dags`
+- use `add_dependency` to avoid deprecation warnings from CDK
+- various typo fixes
+- various clean-ups to the SageMaker Service Catalog templates
+- fix opensearch removal policy
 
 ## v1.2.0